discord_bot/requirements.txt
discord.py
aiohttp
python-dotenv
libsql-client
infisical-python
python-dateutil
toml
requests
beautifulsoup4
